modified:   Assembly/AssemblyAlgorithms.cpp 	modified:   Common/Functional.h 	modified:   Common/StringUtil.h 	modified:   Graph/AdjIO.h 	modified:   Graph/ContigGraphAlgorithms.h 	modified:   ParseAligns/ParseAligns.cpp 	modified:   ParseAligns/abyss-fixmate.cc 	modified:   Scaffold/scaffold.cc 	modified:   configure.ac

+++ Assembly/AssemblyAlgorithms.cpp
@@ -962,7 +962,7 @@ static float calculateCoverageThreshold(const Histogram& h)
 			logger(1) << "Coverage: " << cov << "\t"
 				"Reconstruction: " << trimmed.size() << endl;
 
-		unsigned median = trimmed.median();
+		double median = trimmed.median();
 		float cov1 = sqrt(median);
 		if (cov1 == cov) {
 			// The coverage threshold has converged.
